+++ migrations/env.py
@@ -1,12 +1,12 @@
-from __future__ import with_statement
-
 import logging
 from logging.config import fileConfig
 
-from sqlalchemy import engine_from_config
-from sqlalchemy import pool
+from flask import current_app
 
 from alembic import context
+
+# do NOT use foreign key checking when updating DB
+current_app.config["SQLITE_FOREIGN_KEYS"] = False
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
@@ -17,25 +17,40 @@ config = context.config
 fileConfig(config.config_file_name)
 logger = logging.getLogger("alembic.env")
 
+
+def get_engine():
+    try:
+        # this works with Flask-SQLAlchemy<3 and Alchemical
+        return current_app.extensions["migrate"].db.get_engine()
+    except TypeError:
+        # this works with Flask-SQLAlchemy>=3
+        return current_app.extensions["migrate"].db.engine
+
+
+def get_engine_url():
+    try:
+        return get_engine().url.render_as_string(hide_password=False).replace("%", "%%")
+    except AttributeError:
+        return str(get_engine().url).replace("%", "%%")
+
+
 # add your model's MetaData object here
 # for 'autogenerate' support
 # from myapp import mymodel
 # target_metadata = mymodel.Base.metadata
-from flask import current_app
-
-current_app.config[
-    "SQLITE_FOREIGN_KEYS"
-] = False  # do NOT use foreign key checking when updating DB
-config.set_main_option(
-    "sqlalchemy.url",
-    str(current_app.extensions["migrate"].db.engine.url).replace("%", "%%"),
-)
-target_metadata = current_app.extensions["migrate"].db.metadata
+config.set_main_option("sqlalchemy.url", get_engine_url())
+target_db = current_app.extensions["migrate"].db
 
 # other values from the config, defined by the needs of env.py,
 # can be acquired:
 # my_important_option = config.get_main_option("my_important_option")
 # ... etc.
+
+
+def get_metadata():
+    if hasattr(target_db, "metadatas"):
+        return target_db.metadatas[None]
+    return target_db.metadata
 
 
 def run_migrations_offline():
@@ -51,8 +66,9 @@ def run_migrations_offline():
 
     """
     url = config.get_main_option("sqlalchemy.url")
+    # rencer as batch for sqlite support
     context.configure(
-        url=url, target_metadata=target_metadata, literal_binds=True, render_as_batch=True
+        url=url, target_metadata=get_metadata(), literal_binds=True, render_as_batch=True
     )
 
     with context.begin_transaction():
@@ -77,20 +93,15 @@ def run_migrations_online():
                 directives[:] = []
                 logger.info("No changes in schema detected.")
 
-    connectable = engine_from_config(
-        config.get_section(config.config_ini_section),
-        prefix="sqlalchemy.",
-        poolclass=pool.NullPool,
-    )
+    connectable = get_engine()
 
     with connectable.connect() as connection:
         context.configure(
             connection=connection,
-            target_metadata=target_metadata,
+            target_metadata=get_metadata(),
             process_revision_directives=process_revision_directives,
-            render_as_batch=True,
-            compare_type=True,
-            **current_app.extensions["migrate"].configure_args
+            # compare type and render as batch are set by extension settings now!
+            **current_app.extensions["migrate"].configure_args,
         )
 
         with context.begin_transaction():